[renderer] merging topic manager into minutes store
