+++ src/renderer/views/component/topic/TopicsElement.tsx
@@ -27,12 +27,12 @@ import {
 import { Button, Chip } from "@mui/material";
 import { useEffect, useState } from "react";
 import { secondsToHMS } from "../../../util.js";
+import { Topic } from "../../../../common/content/topic.js";
 import { DetailViewDialogState } from "../common/useDetailViewDialog.jsx";
 import { useDiscussionHistory } from "../discussion/DiscussionHistory.jsx";
 import { useMinutesAgendaStore } from "../../store/useAgendaStore.jsx";
-import { useTopicStore } from "../../store/useTopicManagerStore.jsx";
+import { useTopicStore } from "../../store/useMinutesStore.jsx";
 import { useVBStore } from "../../store/useVBStore.jsx";
-import { Topic } from "../../../../common/content/topic.js";
 import { TopicAIConfigDialog } from "./TopicAIConfigDialog.jsx";
 import { useMinutesStore } from "../../store/useMinutesStore.jsx";
 import { processTopicAction } from "../../action/TopicAction.js";
